preliminary profile CRUD flow by admin

+++ models/dao/interfaces/userdatabase.go
@@ -4,10 +4,10 @@ import "go-team-room/models/dao"
 
 type UserDatabase interface {
   AddUser(user *dao.User) (int64, error)
-  DeleteUser(id int) error
-  UpdateUser(id int, user *dao.User) error
+  DeleteUser(id int64) error
+  UpdateUser(id int64, user *dao.User) error
 
-  FindUserById(id int) (*dao.User, error)
+  FindUserById(id int64) (*dao.User, error)
   FindUserByEmail(email string) (*dao.User, error)
   FindUserByPhone(phone string) (*dao.User, error)
 }
